Display Current Appointments
Display Current Appointments 
This script will display the current month’s calendar in a web page (HTML). It will then look at all of the  appointments in the calendar, and display the ones that are current in the current week, the next week, and the  week after that. 
The program has set a switch to look at ALL occurrences of recurring tasks. This can amount to a high number,  and will be an unlimited number if no end date is specified. Therefore, the program MUST provide a limit to the  number of recurring items it looks at for each appointment. This one looks at the first 365, so it will see an item  that started as far back as a year. 
Features: 
∙ Create a monthly calendar 
∙ Create a web page 
∙ Interview all appointments in the calendar